backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
motor==3.3.1
pymongo==4.6.0
bleach==6.1.0
python-json-logger==2.0.7
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
websockets==12.0
grpcio==1.59.0
grpcio-tools==1.59.0
jinja2==3.1.2
reportlab==4.0.7
python-multipart==0.0.6
requests==2.31.0